+++ libcoveoc4ids/additional_checks.py
@@ -1,52 +1,38 @@
 import gettext
 import re
 
+ORG_PATHS = re.compile(
+    r"""
+    ^/projects/(\d+)/(
+        publicAuthority|
+        budget/budgetBreakdowns/\d+/budgetBreakdown/\d+/sourceParty|
+        contractingProcesses/\d+/summary/(
+            tender/(
+                tenderers/\d+|
+                procuringEntity|
+                administrativeEntity
+            )|
+            suppliers/\d+
+        )
+    )/id$
+    """,
+    re.VERBOSE,
+)
 
-class CurrencyCheck:
-    def process(self, data, flat_data):
-        amount_path = re.compile(".*amount$")
-
-        expected_currency_paths = []
-        missing = []
-
-        for path in flat_data.keys():
-            if amount_path.match(path):
-                expected_currency_paths.append(path[: -len("amount")] + "currency")
-
-        for expected in expected_currency_paths:
-            if expected not in flat_data:
-                missing.append(expected[: -len("currency")])
-
-        if len(missing) == 0:
-            return True
-
-        return {
-            "check_id": "missing-currency",
-            "message": gettext.gettext(
-                "There are %(count)d values without a currency. Currencies should be published for all values."
-            )
-            % {"count": len(missing)},
-            "paths": missing,
-        }
+# matches oc4ids-abc123-anything
+PROJECT_ID = re.compile(r"^oc4ids\-[a-z0-9]{6}.")
 
 
-class EmptyValueCheck:
-    def process(self, data, flat_data):
-        missing = []
+# See libcoveocds' empty_field.
+def empty_value(data, flat):
+    missing = [
+        path
+        for path, value in flat.items()
+        if isinstance(value, str) and not value.strip() or isinstance(value, (dict, list)) and not value
+    ]
 
-        # Similar check to libcoveocds EmptyFieldCheck
-        for path, value in flat_data.items():
-            if isinstance(value, str) and len(value.strip()) == 0:
-                missing.append(path)
-            elif isinstance(value, dict) and len(value) == 0:
-                missing.append(path)
-            elif isinstance(value, list) and len(value) == 0:
-                missing.append(path)
-
-        if len(missing) == 0:
-            return True
-
-        return {
+    if missing:
+        yield {
             "check_id": "missing-values",
             "message": gettext.gettext(
                 "The data includes fields that are empty or contain only whitespaces. "
@@ -57,64 +43,75 @@ class EmptyValueCheck:
         }
 
 
-class OrgReferencesExistCheck:
-    project_id_match = re.compile(r"^/projects/\d+")
+def currency(data, flat):
+    missing = [
+        path[:-6] for path in flat if path.endswith("amount") and f"{path[:-6]}currency" not in flat  # len("amount")
+    ]
 
-    def __init__(self, *args, **kwargs):
-        # Cache of project party ids
-        self._cached_project_parties_ids = {}
+    if missing:
+        yield {
+            "check_id": "missing-currency",
+            "message": gettext.gettext(
+                "There are %(count)d values without a currency. Currencies should be published for all values."
+            )
+            % {"count": len(missing)},
+            "paths": missing,
+        }
 
-    def extract_project_from_path(self, path):
-        """Matches /projects/<int> and returns the int portion"""
 
-        match = self.project_id_match.match(path)
-        if match:
-            return int(match.group()[len("/projects/") :])
+def org_references_exist(data, flat):
+    missing = []
 
-    def project_parties_ids(self, data, project):
-        try:
-            return self._cached_project_parties_ids[project]
-        except KeyError:
-            try:
-                self._cached_project_parties_ids[project] = [
-                    party["id"] for party in data["projects"][project]["parties"]
-                ]
-                return self._cached_project_parties_ids[project]
-            except KeyError:
-                # This happens if the project has no parties defined
-                return []
+    parties_ids = {}
 
-    def process(self, data, flat_data):
-        org_paths = re.compile(
-            r"(.*)(publicAuthority|budget\/budgetBreakdown\/\d+\/sourceParty|"
-            r"contractingProcesses\/\d+\/summary\/(tender|suppliers\/\d+)(\/)"
-            r"*(tenderers\/\d+|procuringEntity|administrativeEntity)*)\/id$"
-        )
+    for path, value in flat.items():
+        if match := ORG_PATHS.search(path):
+            project_index = int(match.group(1))
 
-        missing_references_paths = []
+            if project_index not in parties_ids:
+                try:
+                    parties_ids[project_index] = [party["id"] for party in data["projects"][project_index]["parties"]]
+                except KeyError:
+                    parties_ids[project_index] = []
 
-        for path, value in flat_data.items():
-            if org_paths.match(path):
-                project = self.extract_project_from_path(path)
-                if value not in self.project_parties_ids(data, project):
-                    missing_references_paths.append(path)
+            if value not in parties_ids[project_index]:
+                missing.append(path)
 
-        if len(missing_references_paths) == 0:
-            return True
-
-        return {
+    if missing:
+        yield {
             "check_id": "missing-org-refs",
             "message": gettext.gettext(
                 "There are %(count)d organization references with an id that does not match the id of any parties. "
                 "All organization references should have an associated entry in the parties array with a matching id."
             )
-            % {"count": len(missing_references_paths)},
-            "paths": missing_references_paths,
+            % {"count": len(missing)},
+            "paths": missing,
         }
 
 
-additional_checks = [
-    EmptyValueCheck(),
-    CurrencyCheck(),
-    OrgReferencesExistCheck(),
-]
+def project_prefix(data, flat):
+    invalid = []
+
+    if "projects" in data:
+        projects = data["projects"]
+        if isinstance(projects, list):
+            for i, project in enumerate(projects):
+                if "id" in project:
+                    project_id = project["id"]
+                    if isinstance(project_id, str) and not PROJECT_ID.search(project_id):
+                        invalid.append({"path": f"/projects/{i}/id", "value": project["id"]})
+
+    if invalid:
+        yield {
+            "check_id": "invalid-project-ids",
+            "message": gettext.gettext(
+                "%(count)d of your project id fields has a problem: "
+                "There is no prefix or the prefix format is not recognised."
+            )
+            % {"count": len(invalid)},
+            "path_values": invalid,
+        }
+
+
+ADDITIONAL_CHECKS = [empty_value, currency, org_references_exist]
+CONFORMANCE_CHECKS = [project_prefix]